requirements.txt
deepgram-sdk
requests
python-dotenv
flask
groq
